requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cloudinary==1.29.0
Django==3.2.14
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
phonenumbers==8.12.51
Pillow==9.2.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
